19059 reorganize for new health api (#5949)
* Creating shared directory for common code

* Create health_api folder and move patient inside

* Moving fhir_client.rb into the lighthouse folder

* Move fhir_headers.rb inside lighthouse
