+++ modules/health_quest/app/controllers/health_quest/v0/patients_controller.rb
@@ -14,7 +14,7 @@ module HealthQuest
       private
 
       def factory
-        @factory ||= PatientGeneratedData::Patient::Factory.manufacture(current_user)
+        @factory ||= HealthApi::Patient::Factory.manufacture(current_user)
       end
     end
   end
