+++ modules/health_quest/app/services/health_quest/shared/identity_meta_info.rb
@@ -1,0 +1,107 @@
+# frozen_string_literal: true
+
+module HealthQuest
+  module Shared
+    module IdentityMetaInfo
+      ##
+      # The namespace URI for the FHIR::Identifier object
+      #
+      SYSTEM_ID = 'urn:uuid:2.16.840.1.113883.4.349'
+      ##
+      # Identity of the terminology system for the FHIR::Identifier object
+      #
+      CODING_SYSTEM = 'https://pki.dmdc.osd.mil/milconnect'
+      ##
+      # Operation Resource meta tag for the FHIR::Meta object
+      #   VA URI
+      META_SYSTEM = 'https://wiki.mobilehealth.va.gov/display/PGDMS/Client+Provenance+Mapping'
+      ##
+      # Operation Resource meta tag for the FHIR::Meta object
+      #   VA identifier
+      META_CODE = 'vagov-a0e116eb-faa1-4703-aafe-1a270128607a'
+      ##
+      # Operation Resource meta tag for the FHIR::Meta object
+      #   VA application identifier
+      META_DISPLAY = 'VA GOV CLIPBOARD'
+
+      ##
+      # Builds and sets attributes on the FHIR::Identifier object.
+      #
+      # @return [FHIR::Identifier]
+      #
+      def set_identifiers
+        identifier.tap do |i|
+          i.value = identifier_value
+          i.system = SYSTEM_ID
+          i.type = identifier_type
+        end
+      end
+
+      ##
+      # Build the hash for the Patient objects identifier type.
+      #
+      # @return [Hash]
+      #
+      def identifier_type
+        {
+          coding: [{
+            system: CODING_SYSTEM,
+            code: identifier_code,
+            userSelected: false
+          }]
+        }
+      end
+
+      ##
+      # Builds and sets the tag array attribute on the FHIR::Meta object.
+      #
+      # @return [FHIR::Meta]
+      #
+      def set_meta
+        meta.tap do |m|
+          m.tag = [{
+            system: META_SYSTEM,
+            code: META_CODE,
+            display: META_DISPLAY
+          }]
+        end
+      end
+
+      ##
+      # Method non-implementation warning.
+      #
+      # @return [NotImplementedError]
+      #
+      def identifier
+        raise NotImplementedError "#{self.class} should have implemented identifier ..."
+      end
+
+      ##
+      # Method non-implementation warning.
+      #
+      # @return [NotImplementedError]
+      #
+      def meta
+        raise NotImplementedError "#{self.class} should have implemented meta ..."
+      end
+
+      ##
+      # Method non-implementation warning.
+      #
+      # @return [NotImplementedError]
+      #
+      def identifier_value
+        raise NotImplementedError "#{self.class} should have implemented identifier_value ..."
+      end
+
+      ##
+      # Method non-implementation warning.
+      #
+      # @return [NotImplementedError]
+      #
+      def identifier_code
+        raise NotImplementedError "#{self.class} should have implemented identifier_code ..."
+      end
+    end
+  end
+end
