+++ modules/health_quest/spec/request/patients_request_spec.rb
@@ -36,7 +36,7 @@ RSpec.describe 'health_quest patients', type: :request do
       before do
         sign_in_as(current_user)
         allow_any_instance_of(HealthQuest::Lighthouse::Session).to receive(:retrieve).and_return(session_store)
-        allow_any_instance_of(HealthQuest::PatientGeneratedData::Patient::MapQuery)
+        allow_any_instance_of(HealthQuest::HealthApi::Patient::MapQuery)
           .to receive(:get).with(anything).and_return(client_reply)
       end
 
@@ -79,7 +79,7 @@ RSpec.describe 'health_quest patients', type: :request do
       before do
         sign_in_as(current_user)
         allow_any_instance_of(HealthQuest::Lighthouse::Session).to receive(:retrieve).and_return(session_store)
-        allow_any_instance_of(HealthQuest::PatientGeneratedData::Patient::MapQuery)
+        allow_any_instance_of(HealthQuest::HealthApi::Patient::MapQuery)
           .to receive(:create).with(anything).and_return(client_reply)
       end
