+++ modules/health_quest/spec/services/patient_generated_data/questionnaire_response/resource_spec.rb
@@ -27,8 +27,8 @@ describe HealthQuest::PatientGeneratedData::QuestionnaireResponse::Resource do
   end
 
   describe 'included modules' do
-    it 'includes PatientGeneratedData::Common::IdentityMetaInfo' do
-      expect(subject.ancestors).to include(HealthQuest::PatientGeneratedData::Common::IdentityMetaInfo)
+    it 'includes Shared::IdentityMetaInfo' do
+      expect(subject.ancestors).to include(HealthQuest::Shared::IdentityMetaInfo)
     end
   end
